+++ Classes/Controller/AddressController.php
@@ -32,6 +32,7 @@ use SICOR\SicAddress\Domain\Repository\AddressRepository;
 use SICOR\SicAddress\Domain\Repository\CategoryRepository;
 use SICOR\SicAddress\Domain\Repository\ContentRepository;
 use SICOR\SicAddress\Domain\Service\GeocodeService;
+use TYPO3\CMS\Core\Page\AssetCollector;
 use TYPO3\CMS\Core\Pagination\SimplePagination;
 use TYPO3\CMS\Core\Type\ContextualFeedbackSeverity;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -103,11 +104,7 @@ class AddressController extends AbstractController
 
         // Make search respect configured pages if there are some
         $pages = $this->request->getAttribute('currentContentObject')->data['pages'];
-
-        if (
-            isset($pages) &&
-            strlen($pages) > 0
-        ) {
+        if (is_string($pages) && trim($pages) !== '') {
             $this->querySettings->setRespectStoragePage(TRUE);
             $this->querySettings->setStoragePageIds(explode(',', $pages));
         } else {
@@ -117,7 +114,11 @@ class AddressController extends AbstractController
         $this->addressRepository->setDefaultQuerySettings($this->querySettings);
 
         // Include js
-        $GLOBALS['TSFE']->additionalFooterData['tx_sicaddress_sicaddress'] = '<script src="typo3conf/ext/sic_address/Resources/Public/Javascript/sicaddress.js" type="text/javascript"></script>';
+        GeneralUtility::makeInstance(AssetCollector::class)->addJavaScript(
+            'tx_sicaddress_sicaddress',
+            'EXT:sic_address/Resources/Public/Javascript/sicaddress.js',
+            ['defer' => 'defer']
+        );
     }
 
     protected function initializeView($view)
@@ -352,24 +353,22 @@ class AddressController extends AbstractController
             $args['distance'] = null;
         }
 
-        $mapCenter = new \SICOR\SicAddress\Domain\Model\Address();
+        $mapCenter = new Address();
         $centerAddress = $this->service->getCenterAddressObjectFromFlexConfig();
+        $currentCountry = $args['country'] ?? "Deutschland";
+        $centerNotFound = false;
+
         if($centerAddress) {
             // Default: Use coordinates of center address for map center
             $mapCenter = clone $centerAddress;
+        } else {
+            // Use country center if available
+            $this->updateMapCenter($mapCenter, '', $currentCountry);
         }
 
-        $centerNotFound = false;
-        if(!empty($args['center'])) {
-            $currentCountry = $arg['country'] ?? "Deutschland";
-            $searchCenter = $this->geocodeService->getCoordinatesForPostalCode($args['center'], $currentCountry);
-            if($searchCenter && !empty($searchCenter['longitude']) && !empty($searchCenter['latitude'])) {
-                // Search: Use coordinates of found address for map center
-                $mapCenter->setLongitude($searchCenter['longitude']);
-                $mapCenter->setLatitude($searchCenter['latitude']);
-            }
-            else {
-                // We tried, but couldn't find it...
+        if (!empty($args['center'])) {
+            // Try to use center from user input (postal code)
+            if (!$this->updateMapCenter($mapCenter, $args['center'], $currentCountry)) {
                 $centerNotFound = true;
             }
         }
@@ -399,6 +398,21 @@ class AddressController extends AbstractController
             'distances' => $this->getDistances(),
             'radius' => $args['distance'],
         ]);
+    }
+
+    /**
+    * Helper method: calls the geocode service and writes found coordinates into the Address object.
+    * Returns true if coordinates were set, otherwise false.
+    */
+    private function updateMapCenter(Address $mapCenter, $postalcode, $country): bool
+    {
+        $searchCenter = $this->geocodeService->getCoordinatesForPostalCode($postalcode, $country);
+        if ($searchCenter && !empty($searchCenter['longitude']) && !empty($searchCenter['latitude'])) {
+            $mapCenter->setLongitude($searchCenter['longitude']);
+            $mapCenter->setLatitude($searchCenter['latitude']);
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -617,7 +631,6 @@ class AddressController extends AbstractController
             case 'nicosdir': $template = 'NicosList.html'; break;
             case 'diakonie': $template = 'DiakonieList.html'; break;
             case 'duelmen': $template = 'DuelmenList.html'; break;
-            case 'massiv': $template = 'MassivList.html'; break;
             case 'obgdir': $template = 'OBGList.html'; break;
         }
         if (method_exists($this->view, 'setTemplate')) {