+++ Classes/Domain/Model/Category.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace SICOR\SicAddress\Domain\Model;
 
 /***************************************************************
@@ -44,8 +45,6 @@ class Category extends \TYPO3\CMS\Extbase\Domain\Model\Category
 
     /**
      * Initialize images
-     *
-     * @return \GeorgRinger\News\Domain\Model\Category
      */
     public function __construct()
     {
@@ -55,28 +54,32 @@ class Category extends \TYPO3\CMS\Extbase\Domain\Model\Category
     /**
      * @return FileReference[]|\TYPO3\CMS\Extbase\Persistence\ObjectStorage
      */
-    public function getSicAddressMarker() {
+    public function getSicAddressMarker()
+    {
         return $this->sicAddressMarker;
     }
 
     /**
      * @param FileReference $sicAddressMarker
      */
-    public function setSicAddressMarker(FileReference $sicAddressMarker) {
+    public function setSicAddressMarker(FileReference $sicAddressMarker)
+    {
         $this->sicAddressMarker = $sicAddressMarker;
     }
 
     /**
      * @param FileReference $sicAddressMarker
      */
-    public function addSicAddressMarker(FileReference $sicAddressMarker) {
+    public function addSicAddressMarker(FileReference $sicAddressMarker)
+    {
         $this->sicAddressMarker->attach($sicAddressMarker);
     }
 
     /**
      * @param FileReference $sicAddressMarker
      */
-    public function removeSicAddressMarker(FileReference $sicAddressMarker) {
+    public function removeSicAddressMarker(FileReference $sicAddressMarker)
+    {
         $this->sicAddressMarker->detach($sicAddressMarker);
     }