+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.1.1] - 2026-05-15
+
+### Fixed
+- Fixed Category Search.
+
+### Minor
+- Code improvements for better TYPO3 13 compatibility.
+
+## [6.1.0] - 2026-01-27
+
+### Changed
+- Improved map related code in address controller and repository.
+
+### Minor
+- Code improvements as suggested by Franz Holzinger.
+
 ## [6.0.1] - 2025-11-07
 
 ### Minor